Split TokenService into smaller pieces
