--- oauth2-server-core/src/main/java/nl/myndocs/oauth2/TokenService.kt
@@ -1,33 +1,0 @@
-package nl.myndocs.oauth2
-
-import nl.myndocs.oauth2.authenticator.Authenticator
-import nl.myndocs.oauth2.authenticator.IdentityScopeVerifier
-import nl.myndocs.oauth2.identity.TokenInfo
-import nl.myndocs.oauth2.request.*
-import nl.myndocs.oauth2.response.TokenResponse
-import nl.myndocs.oauth2.token.AccessToken
-import nl.myndocs.oauth2.token.CodeToken
-
-interface TokenService {
-    fun authorize(passwordGrantRequest: PasswordGrantRequest): TokenResponse
-
-    fun authorize(authorizationCodeRequest: AuthorizationCodeRequest): TokenResponse
-
-    fun authorize(clientCredentialsRequest: ClientCredentialsRequest): TokenResponse
-
-    fun refresh(refreshTokenRequest: RefreshTokenRequest): TokenResponse
-
-    fun redirect(
-            redirect: RedirectAuthorizationCodeRequest,
-            authenticator: Authenticator?,
-            identityScopeVerifier: IdentityScopeVerifier?
-    ): CodeToken
-
-    fun redirect(
-            redirect: RedirectTokenRequest,
-            authenticator: Authenticator?,
-            identityScopeVerifier: IdentityScopeVerifier?
-    ): AccessToken
-
-    fun tokenInfo(accessToken: String): TokenInfo
-}
