+++ oauth2-server-core/src/main/java/nl/myndocs/oauth2/config/Configuration.kt
@@ -1,12 +1,10 @@
 package nl.myndocs.oauth2.config
 
 import nl.myndocs.oauth2.CallRouter
-import nl.myndocs.oauth2.TokenService
 import nl.myndocs.oauth2.authenticator.Authorizer
 import nl.myndocs.oauth2.request.CallContext
 
 data class Configuration(
-        val tokenService: TokenService,
         val callRouter: CallRouter,
         val authorizerFactory: (CallContext) -> Authorizer
 )
