+++ oauth2-server-core/src/main/java/nl/myndocs/oauth2/token/converter/Converters.kt
@@ -1,0 +1,7 @@
+package nl.myndocs.oauth2.token.converter
+
+data class Converters(
+        val accessTokenConverter: AccessTokenConverter,
+        val refreshTokenConverter: RefreshTokenConverter,
+        val codeTokenConverter: CodeTokenConverter
+)
