+++ oauth2-server-core/src/test/java/nl/myndocs/oauth2/AuthorizationCodeGrantTokenServiceTest.kt
@@ -1,7 +1,6 @@
 package nl.myndocs.oauth2
 
 import io.mockk.every
-import io.mockk.impl.annotations.InjectMockKs
 import io.mockk.impl.annotations.MockK
 import io.mockk.impl.annotations.RelaxedMockK
 import io.mockk.junit5.MockKExtension
@@ -11,23 +10,30 @@ import nl.myndocs.oauth2.client.ClientService
 import nl.myndocs.oauth2.exception.InvalidClientException
 import nl.myndocs.oauth2.exception.InvalidGrantException
 import nl.myndocs.oauth2.exception.InvalidRequestException
+import nl.myndocs.oauth2.grant.GrantingCall
+import nl.myndocs.oauth2.grant.authorize
 import nl.myndocs.oauth2.identity.Identity
 import nl.myndocs.oauth2.identity.IdentityService
 import nl.myndocs.oauth2.request.AuthorizationCodeRequest
+import nl.myndocs.oauth2.request.CallContext
 import nl.myndocs.oauth2.token.AccessToken
 import nl.myndocs.oauth2.token.CodeToken
 import nl.myndocs.oauth2.token.RefreshToken
 import nl.myndocs.oauth2.token.TokenStore
 import nl.myndocs.oauth2.token.converter.AccessTokenConverter
 import nl.myndocs.oauth2.token.converter.CodeTokenConverter
+import nl.myndocs.oauth2.token.converter.Converters
 import nl.myndocs.oauth2.token.converter.RefreshTokenConverter
 import org.junit.jupiter.api.Assertions.assertThrows
+import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import java.time.Instant
 
 @ExtendWith(MockKExtension::class)
 internal class AuthorizationCodeGrantTokenServiceTest {
+    @MockK
+    lateinit var callContext: CallContext
     @MockK
     lateinit var identityService: IdentityService
     @MockK
@@ -41,8 +47,22 @@ internal class AuthorizationCodeGrantTokenServiceTest {
     @MockK
     lateinit var codeTokenConverter: CodeTokenConverter
 
-    @InjectMockKs
-    lateinit var tokenService: Oauth2TokenService
+    lateinit var grantingCall: GrantingCall
+
+    @BeforeEach
+    fun initialize() {
+        grantingCall = object : GrantingCall {
+            override val callContext = this@AuthorizationCodeGrantTokenServiceTest.callContext
+            override val identityService = this@AuthorizationCodeGrantTokenServiceTest.identityService
+            override val clientService = this@AuthorizationCodeGrantTokenServiceTest.clientService
+            override val tokenStore = this@AuthorizationCodeGrantTokenServiceTest.tokenStore
+            override val converters = Converters(
+                    this@AuthorizationCodeGrantTokenServiceTest.accessTokenConverter,
+                    this@AuthorizationCodeGrantTokenServiceTest.refreshTokenConverter,
+                    this@AuthorizationCodeGrantTokenServiceTest.codeTokenConverter
+            )
+        }
+    }
 
     val clientId = "client-foo"
     val clientSecret = "client-bar"
@@ -75,7 +95,7 @@ internal class AuthorizationCodeGrantTokenServiceTest {
         every { refreshTokenConverter.convertToToken(username, clientId, requestScopes) } returns refreshToken
         every { accessTokenConverter.convertToToken(username, clientId, requestScopes, refreshToken) } returns accessToken
 
-        tokenService.authorize(authorizationCodeRequest)
+        grantingCall.authorize(authorizationCodeRequest)
     }
 
     @Test
@@ -84,7 +104,7 @@ internal class AuthorizationCodeGrantTokenServiceTest {
 
         assertThrows(
                 InvalidClientException::class.java
-        ) { tokenService.authorize(authorizationCodeRequest) }
+        ) { grantingCall.authorize(authorizationCodeRequest) }
     }
 
     @Test
@@ -95,7 +115,7 @@ internal class AuthorizationCodeGrantTokenServiceTest {
 
         assertThrows(
                 InvalidClientException::class.java
-        ) { tokenService.authorize(authorizationCodeRequest) }
+        ) { grantingCall.authorize(authorizationCodeRequest) }
     }
 
     @Test
@@ -113,7 +133,7 @@ internal class AuthorizationCodeGrantTokenServiceTest {
 
         assertThrows(
                 InvalidRequestException::class.java
-        ) { tokenService.authorize(authorizationCodeRequest) }
+        ) { grantingCall.authorize(authorizationCodeRequest) }
     }
 
     @Test
@@ -131,7 +151,7 @@ internal class AuthorizationCodeGrantTokenServiceTest {
 
         assertThrows(
                 InvalidRequestException::class.java
-        ) { tokenService.authorize(authorizationCodeRequest) }
+        ) { grantingCall.authorize(authorizationCodeRequest) }
     }
 
     @Test
@@ -153,7 +173,7 @@ internal class AuthorizationCodeGrantTokenServiceTest {
 
         assertThrows(
                 InvalidGrantException::class.java
-        ) { tokenService.authorize(authorizationCodeRequest) }
+        ) { grantingCall.authorize(authorizationCodeRequest) }
     }
 
     @Test
@@ -166,7 +186,7 @@ internal class AuthorizationCodeGrantTokenServiceTest {
 
         assertThrows(
                 InvalidGrantException::class.java
-        ) { tokenService.authorize(authorizationCodeRequest) }
+        ) { grantingCall.authorize(authorizationCodeRequest) }
     }
 
 }
