+++ oauth2-server-core/src/test/java/nl/myndocs/oauth2/RefreshTokenGrantTokenServiceTest.kt
@@ -1,7 +1,6 @@
 package nl.myndocs.oauth2
 
 import io.mockk.every
-import io.mockk.impl.annotations.InjectMockKs
 import io.mockk.impl.annotations.MockK
 import io.mockk.impl.annotations.RelaxedMockK
 import io.mockk.junit5.MockKExtension
@@ -12,22 +11,29 @@ import nl.myndocs.oauth2.client.ClientService
 import nl.myndocs.oauth2.exception.InvalidClientException
 import nl.myndocs.oauth2.exception.InvalidGrantException
 import nl.myndocs.oauth2.exception.InvalidRequestException
+import nl.myndocs.oauth2.grant.GrantingCall
+import nl.myndocs.oauth2.grant.refresh
 import nl.myndocs.oauth2.identity.Identity
 import nl.myndocs.oauth2.identity.IdentityService
+import nl.myndocs.oauth2.request.CallContext
 import nl.myndocs.oauth2.request.RefreshTokenRequest
 import nl.myndocs.oauth2.token.AccessToken
 import nl.myndocs.oauth2.token.RefreshToken
 import nl.myndocs.oauth2.token.TokenStore
 import nl.myndocs.oauth2.token.converter.AccessTokenConverter
 import nl.myndocs.oauth2.token.converter.CodeTokenConverter
+import nl.myndocs.oauth2.token.converter.Converters
 import nl.myndocs.oauth2.token.converter.RefreshTokenConverter
 import org.junit.jupiter.api.Assertions
+import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import java.time.Instant
 
 @ExtendWith(MockKExtension::class)
 internal class RefreshTokenGrantTokenServiceTest {
+    @MockK
+    lateinit var callContext: CallContext
     @MockK
     lateinit var identityService: IdentityService
     @MockK
@@ -41,9 +47,22 @@ internal class RefreshTokenGrantTokenServiceTest {
     @MockK
     lateinit var codeTokenConverter: CodeTokenConverter
 
-    @InjectMockKs
-    lateinit var tokenService: Oauth2TokenService
+    lateinit var grantingCall: GrantingCall
 
+    @BeforeEach
+    fun initialize() {
+        grantingCall = object : GrantingCall {
+            override val callContext = this@RefreshTokenGrantTokenServiceTest.callContext
+            override val identityService = this@RefreshTokenGrantTokenServiceTest.identityService
+            override val clientService = this@RefreshTokenGrantTokenServiceTest.clientService
+            override val tokenStore = this@RefreshTokenGrantTokenServiceTest.tokenStore
+            override val converters = Converters(
+                    this@RefreshTokenGrantTokenServiceTest.accessTokenConverter,
+                    this@RefreshTokenGrantTokenServiceTest.refreshTokenConverter,
+                    this@RefreshTokenGrantTokenServiceTest.codeTokenConverter
+            )
+        }
+    }
     val clientId = "client-foo"
     val clientSecret = "client-bar"
     val refreshToken = "refresh-token"
@@ -72,7 +91,7 @@ internal class RefreshTokenGrantTokenServiceTest {
         every { refreshTokenConverter.convertToToken(token) } returns newRefreshToken
         every { accessTokenConverter.convertToToken(username, clientId, scopes, newRefreshToken) } returns accessToken
 
-        tokenService.refresh(refreshTokenRequest)
+        grantingCall.refresh(refreshTokenRequest)
 
 
         verify { tokenStore.storeAccessToken(accessToken) }
@@ -93,7 +112,7 @@ internal class RefreshTokenGrantTokenServiceTest {
 
         Assertions.assertThrows(
                 InvalidRequestException::class.java
-        ) { tokenService.refresh(refreshTokenRequest) }
+        ) { grantingCall.refresh(refreshTokenRequest) }
     }
 
     @Test
@@ -102,7 +121,7 @@ internal class RefreshTokenGrantTokenServiceTest {
 
         Assertions.assertThrows(
                 InvalidClientException::class.java
-        ) { tokenService.refresh(refreshTokenRequest) }
+        ) { grantingCall.refresh(refreshTokenRequest) }
     }
 
     @Test
@@ -113,7 +132,7 @@ internal class RefreshTokenGrantTokenServiceTest {
 
         Assertions.assertThrows(
                 InvalidClientException::class.java
-        ) { tokenService.refresh(refreshTokenRequest) }
+        ) { grantingCall.refresh(refreshTokenRequest) }
     }
 
     @Test
@@ -127,6 +146,6 @@ internal class RefreshTokenGrantTokenServiceTest {
 
         Assertions.assertThrows(
                 InvalidGrantException::class.java
-        ) { tokenService.refresh(refreshTokenRequest) }
+        ) { grantingCall.refresh(refreshTokenRequest) }
     }
 }
